requirements.txt
argcomplete==3.5.3
cfgv==3.4.0
charset-normalizer==3.4.1
colorama==0.4.6
conventional-commits-check==2.9.0
decli==0.6.2
distlib==0.3.9
distro==1.9.0
filelock==3.18.0
gitdb==4.0.12
GitPython==3.1.44
identify==2.6.9
Jinja2==3.1.6
MarkupSafe==3.0.2
nodeenv==1.9.1
npyscreen==4.10.5
packaging==24.2
platformdirs==4.3.7
pre_commit==4.2.0
prompt_toolkit==3.0.50
PyYAML==6.0.2
questionary==2.1.0
smmap==5.0.2
termcolor==2.5.0
tomlkit==0.13.2
virtualenv==20.30.0
wcwidth==0.2.13
